[refactor][STEP5] helper --> shell directory
